Restricted book addition and removal to admin user

+++ routes/routes.js
@@ -36,6 +36,15 @@ function isUserLoggedIn(req, res) {
     }
 }
 
+function isUserAdmin(req, res) {
+    if (req.session.user) {
+        if (req.session.user.isAdmin) {
+            res.status(200).end()
+        }
+    }
+    res.status(403).end()
+}
+
 function postLogin(req, res) {
     let {email, password} = req.body
     let query = `SELECT * FROM users WHERE email = '${email}';`
@@ -65,6 +74,10 @@ function postLogin(req, res) {
 }
 
 function postAddBook(req, res) {
+    if (!req.session.user || !req.session.user.isAdmin) {
+        res.status(403).end()
+        return
+    }
     let {title, author, isbn, isbn13, date, publisher, language, edition, pages, imgURL} = dataTreat(req.body)
     
     // Ensure given URL is valid, if not, add no URL to DB
@@ -104,8 +117,11 @@ function dataTreat(object) {
 
 
 function getBook(req, res) {
-    let {isbn} = req.body
-
+    let {isbn} = req.params
+    let isAdmin = false
+    if (req.session.user && req.session.user.isAdmin) {
+        isAdmin = true
+    }
     let query = `SELECT * FROM books WHERE isbn = '${isbn}';`
     client.query(query)
         .then( result => {
@@ -113,7 +129,7 @@ function getBook(req, res) {
             if (rowCount === 0) {
                 res.status(404).end() // use the right status code
             } else {
-                res.json(rows[0])
+                res.json({book: rows[0], isAdmin: isAdmin})
             }
         })
         .catch( e => {
@@ -123,6 +139,10 @@ function getBook(req, res) {
 }
 
 function removeBook(req, res) {
+    if (!req.session.user || !req.session.user.isAdmin) {
+        res.status(403).send("You sneaky little hacker")
+        return
+    }
     let {isbn} = req.body
     let query = `DELETE FROM books WHERE isbn='${isbn}';`
 
@@ -169,6 +189,7 @@ module.exports = {
     connectToDB,
     logout,
     isUserLoggedIn,
+    isUserAdmin,
     postLogin,
     postAddBook,
     getBook,
